Created the center plates
Created a top and bottom center plate. Each copter arm will be held by 3
tube clamps, two on the bottom plate, another on the top plate. In
future versions, mounting holes for the controller board will be added
as well as slots for attaching the battery with velcro tape and the ESCs
with cable binders.