Created new top & bottom plates using the template
The newly created template was used to create a new set of center plates.
For now, only the drill holes for mounting the boom clamps were created,
assuming 14.2x23mm clamps for 16mm tubes. The drill holes are 4.1mm in
diameter and suitable for M4 bolts.
Further holes to mount the PCB and slots to attach the battery and the ESC
are supposed to follow.