Added holes for the battery mount
The battery will be mounted on a separate plate that will be attached to
the bottom plate via M3 screws with a spacing between the screws large
enough to accommodate batteries with a width up to 65mm.

Also added a lot of measurement labels on a separate layer for information
purposes.