Added mounting holes for the X configuration
With this additional set of holes, the plate can now be used to mount the
flight control and the receiver in either + or X configuration. The reason
why the receiver mounting holes differ is that the flight control leaves
less space on the side between the booms for the receiver.