Created a PCB to mount the NanoWii and sensors.
This PCB features the mounting holes required to attach it to the center
plate and additional ones to hold the NanoWii flight controller and the
Sparkfun 9DOF stick.

The PCB provides an I2C bus connection for the NanoWii, the BMP085
pressure sensor (Sparkfun breakout version) and the 9DOF stick (also
Sparkfun breakout version). From the 9DOF stick, only the compass will be
used for the time being, thus it might be more feasible to obtain the
magnetometer as a single sensor.