Replaced the 9DOF stick with single HMC5883 breakout
Unfortunately, the 9DOF stick interfered with the NanoWii's own sensors on
the I2C bus. As the HMC5883 magnometer was the only sensor really needed
from the stick, the design was changed to accomodate a single 5883
breakout (from Sparkfun) instead of the 9DOF stick.
The PCB was also changed to double sided and is now intended for
professional manufacturing.