Started working on a simple 35cm frame
This frame is intended for my medium sized quadcopter. It is a simple design with an upper and a lower base plate and four rectangular boom. This is my "all-in-one" drawing and it currently only models the lower baseplate. I will split the file onto the individual parts once the design is more mature.